requirements.txt
fastapi==0.100.0
uvicorn==0.23.2
sqlalchemy==2.0.19
pydantic==2.0.3
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
moviepy==1.0.3
Pillow==10.0.0
python-dotenv==1.0.0
bcrypt==4.0.1
alembic==1.11.1 